Move Heart to a separate component
